First half way stable version  - Optimizes the Timeline widget to not draw several things twice  - Added two new tool (Fade and Strobe) - they do work fine  - Fixed the GP5 parser. There were two major findings. Aligned with the Gutiar Pro Python Library now
There are still some things to do, but it looks much better than a few commits back.

+++ Control_Trackbar_Zoom.cpp
@@ -197,7 +197,7 @@ namespace MIDILightDrawer
 			int Slider_Center = New_X + SLIDER_WIDTH / 2;
 
 			// Find closest step
-			int New_Step = (float)Math::Round((float)(Slider_Center - Track_Start) / Step_Size);
+			int New_Step = (int)((float)Math::Round((float)(Slider_Center - Track_Start) / Step_Size));
 			New_Step = Math::Max(0, Math::Min(New_Step, _Values->Length - 1));
 
 			if (New_Step != _Current_Step) {
